tests/wpt/web-platform-tests/navigation-api/navigation-history-entry/entry-after-detach.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func_done(() => {
    let i_entry = i.contentWindow.navigation.currentEntry;
    assert_true(i_entry.sameDocument);
    assert_not_equals(i_entry.url, null);
    assert_not_equals(i_entry.key, "");
    assert_not_equals(i_entry.id, "");
    i.remove();
    assert_false(i_entry.sameDocument);
    assert_equals(i_entry.url, null);
    assert_equals(i_entry.key, "");
    assert_equals(i_entry.id, "");
  });
}, "NavigationHistoryEntry properties after detach");
</script>
