third_party/blink/web_tests/external/wpt/navigation-api/navigation-history-entry/entry-after-detach-expected.txt

This is a testharness.js-based test.
[FAIL] NavigationHistoryEntry properties after detach
  assert_equals: expected (string) "" but got (object) null
Harness: the test ran to completion.
